+++ src/Calculators/Pcb/TrackCurrentIpc2221A/TrackCurrentIpc2221ACalculator.cs
@@ -211,7 +211,6 @@ namespace NinjaCalc.Calculators.Pcb.TrackCurrentIpc2221A {
             //============================================== FINAL ==========================================//
             //===============================================================================================//
 
-
             this.FindDependenciesAndDependants();
             this.RecalculateAllOutputs();
             this.ValidateAllVariables();